moves event boundary to contexts and renames active record to be more constistent
